requirements.txt
# ── Web framework ───────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ── LangChain stack ──────────────────────────────────────────────────────────
langchain>=0.3.14
langchain-classic>=1.0.0
langchain-openai>=0.3.1
langchain-community>=0.3.14
langchain-chroma>=1.0.0
langchain-text-splitters>=0.3.0

# ── Vector DB ────────────────────────────────────────────────────────────────
chromadb>=1.0.0

# ── Embeddings & reranker ─────────────────────────────────────────────────────
openai>=1.59.9
sentence-transformers>=3.0.0

# ── Tokenisation ─────────────────────────────────────────────────────────────
tiktoken>=0.8.0

# ── Hybrid search (BM25) ─────────────────────────────────────────────────────
rank-bm25>=0.2.2
numpy>=1.24.0

# ── HTML parsing & web crawling ───────────────────────────────────────────────
beautifulsoup4==4.12.3
lxml>=5.3.0
requests>=2.32.0

# ── Async HTTP (Jira client) ──────────────────────────────────────────────────
httpx>=0.27.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic>=2.10.5
